Using hooks in the hoc :)
&& fix styles
