+++ src/components/withFirebase.js
@@ -1,30 +1,30 @@
-import React from "react";
+import React, { useState } from "react";
 import { db } from "../config";
 
 let itemsRef = db.ref("/data");
 
 const withFireBase = WrappedComponent => {
-  const HOC = (props) => {
-    const [ items ] = React.useState(0);
+    const FirebaseConnector = (props) => {
+      const [ data, setData ] = useState([]);
 
-    const fetchData = () => {
-      itemsRef.once("value", snapshot => {
-        let data = snapshot.val();
-        this.setState({ items: Object.values(data) });
-      });
+        const fetchData = () => {
+          itemsRef.once("value", snapshot => {
+            const items = snapshot.val();
+            setData(Object.values(items));
+            return items;
+          });
+        };
+
+        return (
+            <WrappedComponent
+                {...props}
+                fetchData={fetchData}
+                items={data}
+            />
+        );
     };
 
-    return (
-      <WrappedComponent
-        {...props}
-        getFirebaseData={fetchData}
-        items={items}
-      />
-    );
-  };
-
-  return HOC;
+    return FirebaseConnector;
 };
 
 export default withFireBase;
-
