--- src/styles/documentListStyles.js
@@ -1,45 +1,0 @@
-import {Dimensions as ScreenDimensions, StyleSheet} from "react-native";
-import Colors from "../constants/colors";
-import Dimensions from "../constants/dimensions";
-
-let width = ScreenDimensions.get("window").width;
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: Colors.containerBackgroundColor
-  },
-  headerImage: {
-    height: 80,
-    width: width,
-    marginBottom: Dimensions.marginBottom
-  },
-  viewContainer: {
-    backgroundColor: Colors.containerBackgroundColor
-  },
-  typeTextStyle: {
-    fontSize: Dimensions.primaryFontSize,
-    fontWeight: "bold",
-    textAlign: "center"
-  },
-  dateTextStyle: {
-    fontSize: Dimensions.secondaryFontSize
-  },
-  item: {
-    height: 150,
-    backgroundColor: Colors.cardBackgroundColor,
-    alignItems: "center",
-    justifyContent: "center",
-    flex: 1,
-    margin: 1,
-    borderRadius: 3
-  },
-  itemInvisible: {
-    backgroundColor: "transparent"
-  },
-  itemText: {
-    color: Colors.textColor
-  }
-});
-
-export default styles;
